% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACseqTFEA.R
\docType{package}
\name{ATACseqTFEA-package}
\alias{ATACseqTFEA}
\alias{ATACseqTFEA-package}
\title{Transcription Factor Enrichment Analysis for ATAC-seq}
\description{
Assay for Transpose-Accessible Chromatin using sequencing
(ATAC-seq) is a technique to assess genome-wide chromatin accessibility
by probing open chromatin with hyperactive mutant Tn5 Transposase that
inserts sequencing adapters into open regions of the genome.
ATACseqTFEA is an improvement of the current computational method that
detects differential activity of transcription factors (TFs).
ATACseqTFEA not only uses the difference of open region information,
but also (or emphasizes) the difference of TFs footprints
(cutting sites or insertion sites). ATACseqTFEA provides an easy,
rigorous way to broadly assess TF activity changes between two conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jianhong/ATACseqTFEA}
  \item Report bugs at \url{https://github.com/jianhong/ATACseqTFEA/issues}
}

}
\author{
\strong{Maintainer}: Jianhong Ou \email{jianhong.ou@duke.edu} (\href{https://orcid.org/0000-0002-8652-2488}{ORCID})

}
