% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFEAresults.R
\docType{class}
\name{TFEAresults-class}
\alias{TFEAresults-class}
\alias{TFEAresults}
\alias{show,TFEAresults-method}
\alias{as}
\alias{coerce,TFEAresults,data.frame-method}
\alias{$,TFEAresults-method}
\alias{$<-,TFEAresults-method}
\alias{[[,TFEAresults,ANY,ANY-method}
\alias{[[<-,TFEAresults,ANY,ANY-method}
\alias{getEnrichmentScore}
\alias{getEnrichmentScore,TFEAresults-method}
\alias{getBindingSites}
\alias{getBindingSites,TFEAresults-method}
\alias{getMotifID}
\alias{getMotifID,TFEAresults-method}
\title{Class \code{"TFEAresults"}}
\usage{
\S4method{show}{TFEAresults}(object)

TFEAresults(...)

\S4method{$}{TFEAresults}(x, name)

\S4method{$}{TFEAresults}(x, name) <- value

\S4method{[[}{TFEAresults,ANY,ANY}(x, i, j, ..., exact = TRUE)

\S4method{[[}{TFEAresults,ANY,ANY}(x, i, j, ...) <- value

getEnrichmentScore(x)

\S4method{getEnrichmentScore}{TFEAresults}(x)

getBindingSites(x)

\S4method{getBindingSites}{TFEAresults}(x)

getMotifID(x)

\S4method{getMotifID}{TFEAresults}(x)
}
\arguments{
\item{object}{an object of trackStyle}

\item{\dots}{Each argument in \dots becomes an slot in the new
\code{"TFEAresults"}-class.}

\item{x}{TFEAresults object.}

\item{name}{A literal character string or a name (possibly backtick quoted).}

\item{value}{value to replace.}

\item{i, j}{indices specifying elements to extract or replace.}

\item{exact}{see \link[base]{Extract}}
}
\value{
A TFEAresults object.
}
\description{
An object of class \code{"TFEAresults"}
  represents the results of \link{TFEA}.
}
\section{Slots}{

\describe{
\item{\code{enrichmentScore}}{\code{"numeric \link[Matrix]{Matrix}"}, specify the enrichment
score for each transcription factor (TF). Every row represents a TF.
The columns represents the accumulated enrichment score for that rank.}

\item{\code{bindingSites}}{\code{\link[GenomicRanges:GRanges-class]{GenomicRanges}}
object. It is keep same length and order as the columns in enrichmentScore.}

\item{\code{motifID}}{\code{"list"}. The ranks of binding sites for each TF.}

\item{\code{resultsTable}}{\code{"data.frame"}. The data frame contains the
summarized enrichment score, the p-value, and adjuct p-value for each TF.}
}}

\examples{
res <- system.file("extdata", "res.rds", package="ATACseqTFEA")
res
}
\concept{TFEAresults}
